radosgw: major change of rgw * root pool
The earlier versions of this cookbook used a unique 'rgw * root pool'
setting for zone vs zonegroup. This lead to weirdness when calling
radosgw-admin unless you were VERY careful to always include the pool
settings in every call to radosgw-admin. The cookbook itself got it
wrong sometimes.

Newer upstream documentation such as
"Understanding a Multi Site Ceph Gateway Installation" (v13.08.2017) [1]
explicitly suggest using the same RADOS pool for all settings, usually
the default of '.rgw.root'.

To do this, we respect older config variables that might have set the
parameters, but if they are not present we skip them entirely in the
config. For symmetry, realm & period config variables were also
introduced.

It also suggests setting the realm in the radosgw config, which is
similarly implemented.

[1] http://ceph.com/wp-content/uploads/2016/07/Understanding-a-Multi-Site-Ceph-Gateway-Installation-170813.pdf

Signed-off-by: Robin H. Johnson <robbat2@gentoo.org>